backend/app-clinica-veterinaria/requirements.txt
asgiref==3.7.2
Django==4.2.2
djangorestframework==3.14.0
mssql-django==1.3
Pillow==10.0.0
pyodbc==4.0.39
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
